Consolidate test env. code for Gradle plug-ins
