--- gradle-plugins/plugin-base/src/test/java/io/spine/tools/gradle/given/Given.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2018, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.tools.gradle.given;
-
-import org.gradle.api.Action;
-
-/**
- * @author Dmytro Dashenkov
- */
-public class Given {
-
-    @SuppressWarnings("DuplicateStringLiteralInspection") // Different semantics.
-    public static final String JAVA_PLUGIN_ID = "java";
-
-    private Given() {
-    }
-
-    public enum NoOp implements Action<Object> {
-        ACTION;
-
-        @Override
-        public void execute(Object ignored) {
-            // NoOp
-        }
-
-        @SuppressWarnings("unchecked") // No matter, as the action is NoOp.
-        public static <T> Action<T> action() {
-            return ((Action<T>) ACTION);
-        }
-    }
-}
